+++ src/HetrickCV.hpp
@@ -70,6 +70,7 @@ extern Model *modelPhasorGates32;
 extern Model *modelPhasorGates64;
 extern Model *modelPhasorGen;
 extern Model *modelPhasorGeometry;
+extern Model *modelPhasorHumanizer;
 extern Model *modelPhasorOctature;
 extern Model *modelPhasorQuadrature;
 extern Model *modelPhasorRandom;